pyproject.toml
[tool.poetry]
name = "drf-qalampir-uz"
version = "0.1.0"
description = ""
authors = ["ozoda <ozodas531@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_qalampir_uz"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
celery = "^5.2.7"
redis = "^4.5.5"
flower = "^1.2.0"
django-mptt = "^0.14.0"
pillow = "^9.5.0"
django-modeltranslation = "^0.18.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
